move treasury function
