+++ data/adapters/decentralgames.ts
@@ -1,5 +1,20 @@
 import { OrganizationData } from '../types'
-import { getDGTreasury } from '../utils/zerion'
+
+export async function getDGTreasury() {
+  const request = await fetch(
+    `https://api.decentral.games/admin/getTreasuryBalanceHistory/week?`,
+    {
+      method: 'GET',
+      headers: {
+        Accept: 'application/json',
+        'Content-Type': 'application/json',
+      },
+    }
+  );
+
+  const json = await request.json()
+  return json.totalBalanceUSD.graph.slice(-1)[0].secondary
+}
 
 export async function getDGData(): Promise<OrganizationData> {
   const treasuryValue = await getDGTreasury()
